+++ index.jsx
@@ -2,31 +2,22 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-const units = ["B", "KB", "MB", "GB", "TB", "PB"];
-const conversionFactor = {
-  B: 0,
-  KB: 1,
-  MB: 2,
-  GB: 3,
-  TB: 4,
-  PB: 5
-};
-const unitSuffix = {
-  B: "byte",
-  KB: "kilobyte",
-  MB: "megabyte",
-  GB: "gigabyte",
-  TB: "terabyte",
-  PB: "petabyte"
+const units = {
+  B: { conversionFactor: 0, suffix: "byte" },
+  KB: { conversionFactor: 1, suffix: "kilobyte" },
+  MB: { conversionFactor: 2, suffix: "megabyte" },
+  GB: { conversionFactor: 3, suffix: "gigabyte" },
+  TB: { conversionFactor: 4, suffix: "terabyte" },
+  PB: { conversionFactor: 5, suffix: "petabyte" }
 };
 
 const ByteConverter = props => {
   const {
     children,
     hideWarnings,
+    useSI,
     suffix,
     addCommas,
-    useSI,
     inUnit,
     outUnit
   } = props;
@@ -36,6 +27,7 @@ const ByteConverter = props => {
   const isInteger = Number.isInteger(kiddo);
   const isFinite = Number.isFinite(kiddo);
   const isPositive = kiddo > 0;
+  const isNotANumber = Number.isNaN(kiddo);
 
   if (!hideWarnings) {
     if (!isInteger) {
@@ -43,20 +35,27 @@ const ByteConverter = props => {
     }
 
     if (!isFinite) {
-      console.warn("ByteConverter must recieve a finite integer as a child"); // eslint-disable-line no-console
+      console.warn("ByteConverter must recieve a finite integer as a child."); // eslint-disable-line no-console
     }
 
     if (!isPositive) {
-      console.warn("ByteConverter must recieve a positive integer as a child"); // eslint-disable-line no-console
+      console.warn("ByteConverter must recieve a positive integer as a child."); // eslint-disable-line no-console
+    }
+
+    if (isNotANumber) {
+      // eslint-disable-next-line no-console
+      console.warn(
+        "ByteConverter must recieve a number or valid number as a string as a child"
+      );
     }
   }
 
   const InConversion = useSI
-    ? 1000 ** conversionFactor[inUnit]
-    : 1024 ** conversionFactor[inUnit];
+    ? 1000 ** units[inUnit].conversionFactor
+    : 1024 ** units[inUnit].conversionFactor;
   const OutConversion = useSI
-    ? 1000 ** conversionFactor[outUnit]
-    : 1024 ** conversionFactor[outUnit];
+    ? 1000 ** units[outUnit].conversionFactor
+    : 1024 ** units[outUnit].conversionFactor;
 
   let conversion = (kiddo * InConversion) / OutConversion;
 
@@ -65,7 +64,7 @@ const ByteConverter = props => {
   }
 
   const suffixDisplay = suffix
-    ? ` ${unitSuffix[outUnit]}${conversion === 1 ? "" : "s"}`
+    ? ` ${units[outUnit].suffix}${conversion === 1 ? "" : "s"}`
     : null;
 
   return (
@@ -83,8 +82,8 @@ ByteConverter.propTypes = {
   useSI: PropTypes.bool,
   suffix: PropTypes.bool,
   addCommas: PropTypes.bool,
-  inUnit: PropTypes.oneOf(units),
-  outUnit: PropTypes.oneOf(units)
+  inUnit: PropTypes.oneOf(Object.keys(units)),
+  outUnit: PropTypes.oneOf(Object.keys(units))
 };
 
 ByteConverter.defaultProps = {
@@ -92,8 +91,8 @@ ByteConverter.defaultProps = {
   useSI: false,
   suffix: false,
   addCommas: false,
-  inUnit: units[0],
-  outUnit: units[2]
+  inUnit: Object.keys(units)[0],
+  outUnit: Object.keys(units)[2]
 };
 
 export default ByteConverter;