+++ CHANGELOG.md
@@ -19,6 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.0.3] - 2019-01-19
+
+### Added
+
+- warning for not a number
+- more test coverage
+
+### Changed
+
+- simplified units logic
+
 ## [1.0.2] - 2019-01-19
 
 ### Added